チャットのモジュールを統合
