+++ backend/src/app.module.ts
@@ -1,11 +1,13 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
+import { JwtModule } from '@nestjs/jwt';
 import { ScheduleModule } from '@nestjs/schedule';
 import { AuthModule } from './auth/auth.module';
 import { BlocksModule } from './blocks/blocks.module';
 import { ChatMessageModule } from './chat-message/chat-message.module';
 import { ChatRoomMemberModule } from './chat-room-member/chat-room-member.module';
 import { ChatRoomModule } from './chat-room/chat-room.module';
+import { ChatModule } from './chat/chat.module';
 import { DmRoomModule } from './dm-room/dm-room.module';
 import { DmModule } from './dm/dm.module';
 import { FileModule } from './file/file.module';
@@ -33,6 +35,8 @@ import { UsersModule } from './users/users.module';
     DmRoomModule,
     ChatRoomMemberModule,
     ScheduleModule.forRoot(),
+    ChatModule,
+    JwtModule.register({}),
   ],
   controllers: [],
   providers: [],
