+++ backend/src/chat-message/chat-message.module.ts
@@ -1,13 +1,12 @@
 import { Module } from '@nestjs/common';
-import { JwtModule } from '@nestjs/jwt';
 import { AuthModule } from 'src/auth/auth.module';
 import { PrismaModule } from '../prisma/prisma.module';
 import { ChatMessageController } from './chat-message.controller';
-import { ChatMessageGateway } from './chat-message.gateway';
 import { ChatMessageService } from './chat-message.service';
 @Module({
-  imports: [PrismaModule, AuthModule, JwtModule.register({})],
+  imports: [PrismaModule, AuthModule],
   controllers: [ChatMessageController],
-  providers: [ChatMessageService, ChatMessageGateway],
+  providers: [ChatMessageService],
+  exports: [ChatMessageService],
 })
 export class ChatMessageModule {}
