+++ backend/src/chat-room-member/chat-room-member.module.ts
@@ -1,10 +1,8 @@
 import { Module, forwardRef } from '@nestjs/common';
-import { JwtModule } from '@nestjs/jwt';
 import { ScheduleModule } from '@nestjs/schedule';
 import { ChatRoomModule } from 'src/chat-room/chat-room.module';
 import { PrismaModule } from 'src/prisma/prisma.module';
 import { ChatRoomMemberController } from './chat-room-member.controller';
-import { ChatRoomMemberGateway } from './chat-room-member.gateway';
 import { ChatRoomMemberService } from './chat-room-member.service';
 
 @Module({
@@ -12,10 +10,9 @@ import { ChatRoomMemberService } from './chat-room-member.service';
     PrismaModule,
     forwardRef(() => ChatRoomModule),
     ScheduleModule.forRoot(),
-    JwtModule.register({}),
   ],
   controllers: [ChatRoomMemberController],
-  providers: [ChatRoomMemberService, ChatRoomMemberGateway],
+  providers: [ChatRoomMemberService],
   exports: [ChatRoomMemberService],
 })
 export class ChatRoomMemberModule {}
